Move files around for Brandon. Enable hogwild. Correct comments.
